Don’t include hover behaviour in colour mixin
By including the hover behaviour in the colour mixin, it was difficult to create a button that did not have hover behaviour. Though the hover behaviour was included in the colour selection because the amount of colour shift on hover could be different depending on the colour of the button, this was only used on the default button and was probably not worth the added complication.

Instead, we simply add the colour shift on hover to the `menu_bar_item_hover` mixin.