Simplify menu item selected markup requirements
We don't want to require a menu item to have an element inside it to have a selected checkmark. For example, if an element is not clickable but is selected, we may only have a text node inside. We still want that meni item to appear selected so we can remove the wildcard child selector and just put the icon on the menu item instead of the child element.