Fix right aligned group of menu bar items
The margin of the menu bar items was dependent on whether the menu items were left or right aligned. We were missing some rules that made the margins work properly in a right aligned group.

Instead, we make the margin even on both sides to avoid needing to change it depending on alignment. In a group we eliminate the margin between items and the group takes on that margin instead. However, we also have negative margin on the menu bar itself to make up for the margins of the first and last elements.