Improve handling of button text and icon colour
If we have a menu with a non-default colour, e.g. a primary-blue menu, the text and icon colour wouldn't be correct in the pressed or unpressed states respectively.

Instead, we now detect whether the icon should be dark or light and use CSS filters to modify it so it contrasts. In this case, icon colour is only flipped between the original dark colour, and white. Additionally, we now use the `button_colour` mixin to style the menu trigger text colour in the active state so the text is visible on the white background instead of relying on its original colour being a contrasting colour to white.